+++ fedot/core/optimisers/gp_comp/archive.py
@@ -1,0 +1,7 @@
+class SimpleArchive(list):
+    def update(self, x):
+        self.extend((x))
+
+    @property
+    def items(self):
+        return list(self)
